add custom exceptions module and exceptions handling on db level

+++ src/db/crud/news.py
@@ -1,5 +1,6 @@
 from db.connector import get_db
 from db.models import News
+from exceptions import DatabaseNotFoundError
 
 
 def add_news(title: str, content: str):
@@ -12,7 +13,9 @@ def add_news(title: str, content: str):
 
 def get_news_by_id(news_id: int) -> News | None:
     with get_db() as session:
-        return session.query(News).filter_by(id=news_id).first()
+        if not (news := session.query(News).filter_by(id=news_id).first()):
+            raise DatabaseNotFoundError('News with id={} not found'.format(news_id))
+        return news
 
 
 def get_news() -> list[News]:
@@ -21,6 +24,10 @@ def get_news() -> list[News]:
 
 
 def remove_news(news_id: int):
+    # check if news with given id in db
+    if not get_news_by_id(news_id):
+        raise DatabaseNotFoundError('News with id={} not found'.format(news_id))
+    # remove
     with get_db() as session:
         session.query(News).filter_by(id=news_id).delete()
         session.commit()
